+++ inst/tinytest/test-flextable.R
@@ -11,3 +11,18 @@ models[['Logit 2']] <- glm(am ~ hp + disp, mtcars, family = binomial())
 # no error with caption and notes
 tab <- modelsummary(models, "flextable", title = "test title", notes = "test note", stars = TRUE)
 expect_inherits(tab, "flextable")
+
+
+
+# # Issue #761
+# d <- data.frame(x = rnorm(100), y = rnorm(100))
+# res <- list(
+#   lm(y ~ x, data = d),
+#   lm(y ~ x + I(x^2), data = d)
+# )
+# modelsummary(
+#   res,
+#   output = "flextable",
+#   estimate = c(ABC = "estimate"),
+#   shape = term ~ model + statistic
+# )